+++ test/unit/gov/epa/emissions/framework/client/meta/versions/EditVersionsPresenterTest.java
@@ -113,7 +113,7 @@ public class EditVersionsPresenterTest extends EmfMockObjectTestCase {
         Version[] versions = new Version[0];
         InternalSource[] internalSources = new InternalSource[0];
 
-        service.stubs().method("getVersions").with(eq(new Long(dataset.getId()))).will(returnValue(versions));
+        service.stubs().method("getVersions").with(eq(new Integer(dataset.getId()))).will(returnValue(versions));
         EmfSession session = session(null, (DataEditorService) service.proxy(), null);
         EditVersionsPresenter presenter = new EditVersionsPresenter(dataset, session);
         view.expects(once()).method("observe").with(same(presenter));
